refactor: using a base nano contract modal for both create and create token nc
